Address the lower parts of the register
When assigning to a register, constants take up as many bytes as the
size of the register, i.e. assigning to a 32-bit register

mov eax, 60

...is really:

mov eax, 00 00 00 60

That's three bytes of zero padding that we don't need.  We need to set
rax explicitly, because it currently contains the return value of the
write() call (hopefully, 13, but possibly lower or -1).

The Intel spec. says that we can address registers by different names,
depending on the operand size (Volume 1, Section 3.4.1: General-Purpose
Registers), so rax is:

rax: 64-bit
eax: lower 32 bits (32-63)
ax: lower 16 bits (48-63)
ah: upper 8 bits of the lower 16 (48-55)
al: lowest 8 bits (56-63)

Addressing the lower sections of the register takes up fewer bytes in
the instruction and the constant, so the 5-byte 'mov eax, 60' can be
replaced with:

xor eax, eax
mov al, 60

That saves another byte and is stil perfectly safe (xor on the eax has
zeroed the other 32 bits for us).