Take advantage of initial values of registers
Since Linux kernel 2.2.0 (1999), ELF executables have been run with a
consistent register state.  In particular, the following registers start
as zero: rax, rbx, rbc, rdx, rsi, rdi, rbp.

See ELF_PLAT_INIT at:
https://github.com/torvalds/linux/blob/master/arch/x86/include/asm/elf.h#L108

As such, for the write() call, we don't need to zero them before we set
the lower bits to what we want them to be (1 in rax for the write()
syscall number and 13 in rdx for the string length).

This saves three bytes off setting each of those registers.